Changes for Ubuntu 20.04
For now removed most of the letsencrypt stuff, what it installed refused to work on Ubuntu 20.04

Ran manual certbot install and left in place here the symbolic linking of resulting cert files. This needs more investigation but works for the cats deploy for now.

Also updated nginx template to get rid of a deprecation warning that could have been interfering with the cerbot nginx plugin failing (but probably wasn't).